+++ lib/main.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:sauron/screens/login.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:sauron/screens/home/home.dart';
+import 'package:sauron/screens/startup/startup.dart';
+import 'package:sauron/screens/login/login.dart';
 
 void main() {
   runApp(SauronApp());
@@ -13,8 +16,13 @@ class SauronApp extends StatelessWidget {
       title: 'Flutter Demo',
       theme: ThemeData(
         primarySwatch: Colors.blue,
+        fontFamily: GoogleFonts.montserrat().fontFamily
       ),
-      home: LoginScreen(),
+      routes: {
+        "/home": (_) => new HomeScreen(),
+        "/login": (_) => new LoginScreen(),
+      },
+      home: HomeScreen(),
     );
   }
 }
